rename typo on dashboard page, move dashboard page file to pages folder, add poppins font in app.css
