+++ src/component/dashboard/financial-chart.js
